+++ CHANGELOG.md
@@ -1,2 +1,5 @@
+## Unreleased
+- Deletes the `json` property before calling `fetch`.
+
 ## 0.1.337 (Dec 11, 2015)
 - Scoped release for personal use and profit.